created a dgi macro that expands to a call to document.getElementById.
also fixed some stuff in
the bubblescript.js to
make this work.

This is mostly just testing
an idea, but i think the
short hand can really make
the code more terse.

While the macro is working
on my end, i'm pretty sure
it needs to be optimized 
so that it is only expanded
one time even if it is
reffered to multiple times.
That is my thinking anyway.